+++ src/containers/home/climate/index.js
@@ -5,15 +5,15 @@ import ClimateController from './controller';
 
 const Climate = () => (
   <View style={styles.container}>
-    <ClimateMain />
-    <ClimateController />
+    <ClimateMain styles={{ flex: 4 }} />
+    <ClimateController styles={{ flex: 1 }} />
   </View>
 );
 
 const styles = StyleSheet.create({
   container: {
-    flex: 1,
-    justifyContent: 'space-between',
+    // flex: 1,
+    // justifyContent: 'space-between',
     backgroundColor: '#F8F8F8',
   },
 });