+++ src/containers/settings/zones/index.js
@@ -1,5 +1,8 @@
 // главная страница для отображения зон
+import PropTypes from 'prop-types';
 import React from 'react';
+import { connect } from 'react-redux';
+import { TYSdk } from 'tuya-panel-kit';
 import { View, StyleSheet, Text, SafeAreaView } from 'react-native';
 import { FontAwesomeIcon } from '@fortawesome/react-native-fontawesome';
 import { faMapMarkerAlt } from '@fortawesome/free-solid-svg-icons';
@@ -7,23 +10,21 @@ import Strings from '../../../i18n';
 import ZoneIScene from './zone1';
 import ZoneIIScene from './zone2';
 
-const zone1 = Strings.getLang('zone1');
-const zone2 = Strings.getLang('zone2');
-
 const ZonesScene = () => (
   <View style={styles.container}>
     <SafeAreaView style={styles.area}>
       <FontAwesomeIcon icon={faMapMarkerAlt} color="#ffb700" size={25} />
       <FontAwesomeIcon icon={faMapMarkerAlt} color="#ff7300" size={20} />
-      <Text style={styles.title}>{zone1}</Text>
+      <Text style={styles.title}>{Strings.getLang('zone1')}</Text>
     </SafeAreaView>
     <ZoneIScene />
     <SafeAreaView style={styles.area}>
       <FontAwesomeIcon icon={faMapMarkerAlt} color="#ffb700" size={20} />
       <FontAwesomeIcon icon={faMapMarkerAlt} color="#ff7300" size={25} />
-      <Text style={styles.title}>{zone2}</Text>
+      <Text style={styles.title}>{Strings.getLang('zone2')}</Text>
     </SafeAreaView>
     <ZoneIIScene />
+    {/* <Text style={styles.warn}>{Strings.getLang('cliwarn')}</Text> */}
   </View>
 );
 
@@ -48,6 +49,16 @@ const styles = StyleSheet.create({
     justifyContent: 'center',
     marginLeft: 10,
   },
+  // warn: {
+  //   textAlign: 'center',
+  //   fontWeight: '400',
+  //   fontSize: 18,
+  //   color: '#FF2D00',
+  //   flexDirection: 'row',
+  //   justifyContent: 'center',
+  //   padding: 80,
+  //   alignSelf: 'auto',
+  // },
 });
 
 export default ZonesScene;